Güncelleme popup ekranında, id text disable yapıldı.

+++ src/main/java/com/userpanel/demo/repository/UserRepository.java
@@ -1,9 +1,11 @@
 package com.userpanel.demo.repository;
 
-import com.userpanel.demo.model.User;
+import com.userpanel.demo.entity.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
+import java.math.BigInteger;
 
-public interface UserRepository extends MongoRepository<User, Integer> {
+
+public interface UserRepository extends MongoRepository<User, BigInteger> {
 
 }
